requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
chromadb==1.5.5
google-genai==1.66.0
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.32.5
beautifulsoup4==4.14.3
streamlit==1.55.0
openpyxl==3.1.5
